+++ include/CeLux/backends/cpu/Decoder.hpp
@@ -8,9 +8,9 @@ namespace celux::backends::cpu
 class Decoder : public celux::Decoder
 {
   public:
-    Decoder(const std::string& filePath,
-            int numThreads)
-        : celux::Decoder( numThreads)
+    Decoder(const std::string& filePath, int numThreads,
+            std::vector<std::shared_ptr<Filter>> filters)
+        : celux::Decoder( numThreads, filters)
     {
         isHwAccel = false;
         initialize(filePath);