+++ include/CeLux/backends/gpu/cuda/Decoder.hpp
@@ -8,8 +8,9 @@ namespace celux::backends::gpu::cuda
 class Decoder : public celux::Decoder
 {
   public:
-    Decoder(const std::string& filePath, int numThreads)
-        : celux::Decoder(numThreads)
+    Decoder(const std::string& filePath, int numThreads,
+            std::vector<std::shared_ptr<Filter>> filters)
+        : celux::Decoder(numThreads, filters)
     {
         isHwAccel = true;
         initialize(filePath);