+++ src/CeLux/python/Bindings.cpp
@@ -10,12 +10,12 @@ PYBIND11_MODULE(celux, m)
     // VideoReader bindings
 
     py::class_<VideoReader>(m, "VideoReader")
-        .def(py::init<const std::string&, int, const std::string&
+        .def(py::init<const std::string&, int, const std::string&, std::vector<std::tuple<std::string, std::string>>
                      >(),
              py::arg("input_path"),
              py::arg("num_threads") =
                  static_cast<int>(std::thread::hardware_concurrency() / 2),
-             py::arg("device") = "cuda")
+             py::arg("device") = "cuda", py::arg("filters") = std::vector<std::tuple<std::string, std::string>>())
         .def("read_frame", &VideoReader::readFrame)
         .def("seek", &VideoReader::seek)
         .def("supported_codecs", &VideoReader::supportedCodecs)