+++ tests/test2.py
@@ -15,8 +15,6 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 import celux_cuda as celux
 celux.set_log_level(celux.LogLevel.off)
 
-STREAM = torch.cuda.Stream("cuda")
-
 from requests.exceptions import RequestException
 
 # Nicer prints
@@ -69,7 +67,7 @@ def processVideoCuda(videoPath):
     try:
         frameCount = 0
         start = time.time()
-        with celux.VideoReader(videoPath, device = "cuda", num_threads=16, stream = STREAM) as reader:
+        with celux.VideoReader(videoPath, device = "cuda", num_threads=16) as reader:
             for frame in reader:
                 if frameCount == 0:
                     logging.info(